requirements.txt
asgiref==3.4.1
Django==3.2.8
django-filter==21.1
djangorestframework==3.11.2
Markdown==3.3.4
pytz==2021.3
sqlparse==0.4.2
psycopg2-binary==2.9.1
python-dotenv==0.19.1
django-cors-headers==3.10.0
djoser==2.0.5
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
Pillow==8.4.0
pandas==1.3.4
numpy==1.17.3